% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_corr_chr.R
\name{combine_corr_chr}
\alias{combine_corr_chr}
\title{Calculate Correlation in Local Ancestry (K = 3)}
\usage{
combine_corr_chr(cor.list)
}
\arguments{
\item{cor.list}{list of data tables, one per chromosome, generated by get_corr_chr().}
}
\value{
A data table with the average correlation in local ancestry, to be used for estimating the number of generations since admixture using \code{\link[STEAM]{get_g}}.
}
\description{
Create single data frame with average correlation of local ancestry
for pairs of SNPs varying distances apart. Combines data for each chromosome
generated by get_corr_chr(). Code is currently applicable
only to admixed populations with three ancestral populations.
}
\seealso{
\code{\link[STEAM]{get_g}} and \code{\link[STEAM]{get_corr_chr}}
}
