% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_correction.R
\name{createSpatialPlotTheme}
\alias{createSpatialPlotTheme}
\title{Create Standard Plot Theme}
\usage{
createSpatialPlotTheme()
}
\value{
ggplot2 theme object
}
\description{
Creates a consistent theme for spatial plots
}
