% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.split.R
\name{data.split}
\alias{data.split}
\title{Data Splitting for Training and Testing Sets}
\usage{
data.split(STEAM.obj, train.ratio)
}
\arguments{
\item{STEAM.obj}{Object of class STEAM.Object}

\item{train.ratio}{A numeric value between 0 and 1 specifying the proportion of the data to include in the training set.}
}
\description{
This function splits the data in a STEAM object into training and testing sets based on a specified ratio.
}
