% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLPFC.R
\docType{data}
\name{DLPFC}
\alias{DLPFC}
\alias{dlpfc}
\title{LIBD Human DLPFC Spatial Transcriptomics Data (Sample 151673)}
\format{
A list with the following elements:
\describe{
\item{spatial}{A data frame of cell spatial coordinates.}
\item{count_exp}{A gene expression count matrix, with genes as rows and cells as columns.}
\item{labels}{A factor or character vector of ground-truth labels.}
}
}
\source{
LIBD: \url{http://research.libd.org/spatialLIBD/}
}
\usage{
data(DLPFC)
}
\description{
A sample dataset containing human dorsolateral prefrontal cortex (DLPFC)
spatial transcriptomics data from the LIBD project, generated with the
10x Genomics Visium platform (sample ID 151673).
}
\details{
This dataset is included as an example to demonstrate STEAM workflows.
It is not required for package functionality.
}
\keyword{datasets}
