% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_ab.R
\name{example_ab}
\alias{example_ab}
\title{Example a,b scalars}
\format{
List with two elements called 'a' and 'b'.
  Each element contains a vector of length m
  (m = no. markers).
}
\source{
Toy example generated with command
  \code{example_ab <- get_ab(get_deltas(example_map), g = 6)}
}
\description{
Vectors of length m which depend on the distances
between markers and the number of generations sincea
admixture.
}
