% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_L.R
\name{example_L}
\alias{example_L}
\title{Example L matrix}
\format{
Matrix with K rows (K = no. ancestral pop)
  and K-1 columns.
}
\source{
Toy example generated with command
  \code{get_L(apply(example_props,2,mean))}
}
\description{
K x K-1 matrix representing square root of
correlation matrix at a single locus.
}
