% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_corr_chr.R
\name{get_corr_chr}
\alias{get_corr_chr}
\title{Calculate Correlation in Local Ancestry (K = 3)}
\usage{
get_corr_chr(chrom, binsize = 0.5, map, pop1.gds, pop2.gds, pop3.gds,
  verbose = TRUE)
}
\arguments{
\item{chrom}{chromosome number that you are analyzing}

\item{binsize}{size (in cM) of distance bins for calculating correlation; default = 0.5 cM}

\item{map}{map file; data frame with, at minimum, columns 'chr' and 'cM'}

\item{pop1.gds}{name of GDS file storing local ancestry calls for chrom of interest, with alleles coded as 1 = pop 1 ancestry, 0 = pop 2 or 3}

\item{pop2.gds}{name of GDS file storing local ancestry calls for chrom of interest, with alleles coded as 1 = pop 2 ancestry, 0 = pop 1 or 3}

\item{pop3.gds}{name of GDS file storing local ancestry calls for chrom of interest, with alleles coded as 1 = pop 3 ancestry, 0 = pop 1 or 2}

\item{verbose}{do you want to print updates to screen; default = TRUE}
}
\value{
A data table with the observed correlation in local ancestry vectors for a subset of loci on this chromosome.
}
\description{
Calculate the correlation of local ancestry vectors
for a single chromosome. Code is currently only applicable to
admixed populations with three ancestral populations.
}
\seealso{
\code{\link[STEAM]{get_g}} and \code{\link[STEAM]{combine_corr_chr}}
}
