% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_thresh_analytic.R
\name{get_thresh_analytic}
\alias{get_thresh_analytic}
\title{Significance Threshold Estimation for Admixture Mapping using Analytic Approximation}
\usage{
get_thresh_analytic(g, delt, L, chr = 22, alpha = 0.05,
  type = "pval", searchint = c(1.96, 8))
}
\arguments{
\item{g}{the generations since admixture.}

\item{delt}{the average spacing (cM per marker) between markers.}

\item{L}{the total length (cM) of the region.}

\item{chr}{the number of chromosomes; default is 22.}

\item{alpha}{the level for family-wise error rate control; default is 0.05.}

\item{type}{the type of threshold that should be returned: \code{"stat"} for test statistic or \code{"pval"} for p-value; defaults to pval.}

\item{searchint}{the range of test statistic thresholds for \code{uniroot} to search; defaults to 1.96--8, corresponding to p-value threshold of 0.05--1.2e-15}
}
\value{
A single nummber indicating the estimated significance threshold (either test statistic or p-value).
}
\description{
Compute genome-wide significance (test statistic or p-value) threshold for admixture
mapping using analytic approximation to the family-wise error rate.
}
\examples{
get_thresh_analytic(g=6, delt=0.2, L=3500) # get p-value threshold
get_thresh_analytic(g=6, delt=0.2, L=3500, type="stat") # get test statistic threshold

}
\seealso{
\code{\link[stats]{uniroot}} for finding roots of functions
}
