% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_thresh_simstat.R
\name{get_thresh_simstat}
\alias{get_thresh_simstat}
\title{Significance Threshold Estimation for Admixture Mapping using Test Statistic Simulation}
\usage{
get_thresh_simstat(g, map, props, nreps = 10000, alpha = 0.05,
  type = "pval")
}
\arguments{
\item{g}{the generations since admixture.}

\item{map}{data frame with m rows and at least 2 columns ('chr' containing chromosome number and 'cM' containing genetic position in centimorgans), where m = no. markers}

\item{props}{data frame (n x K) of admixture proportions, where n = no. individuals and K = no. ancestral populations}

\item{nreps}{the number of repetitions for the simulation study; default is 10000.}

\item{alpha}{the level for family-wise error rate control; default is 0.05.}

\item{type}{the type of threshold that should be returned: \code{"stat"} for test statistic or \code{"pval"} for p-value; defaults to pval.}
}
\value{
A single nummber indicating the estimated significance threshold (either test statistic or p-value).
}
\description{
Compute genome-wide significance (test statistic or p-value) threshold for admixture
mapping by simulating admixture mapping test statistics from their joint asymptotic
distribution under the null (see Grinde et al., 2018) and finding
the threshold that controls the empirical family-wise error rate.
}
\examples{
get_thresh_simstat(g = 6, map = example_map, props = example_props, nreps = 10)
get_thresh_simstat(g = 6, map = example_map, props = example_props, nreps = 10, type="stat")

}
