% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_correction.R
\name{loadPatchwork}
\alias{loadPatchwork}
\title{Load Patchwork Package Safely}
\usage{
loadPatchwork()
}
\value{
TRUE if loaded successfully
}
\description{
Centralized function to load patchwork for combined plots
}
