% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STEAM_Obj.R
\name{LoadSTEAM}
\alias{LoadSTEAM}
\title{STEAM object - Loading data for STEAM}
\usage{
LoadSTEAM(count_exp = NULL, spatial = NULL, labels = NULL, Seurat.obj = NULL)
}
\arguments{
\item{count_exp}{Count matrix/ Gene expression data}

\item{spatial}{Spatial Coordinates}

\item{labels}{Cluster info/annotations}

\item{Seurat.obj}{SEURAT Obj}
}
\value{
Object of class 'STEAM.Object'
}
\description{
STEAM object - Loading data for STEAM
}
