% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.train.R
\name{model.train}
\alias{model.train}
\title{Train a Classification Model}
\usage{
model.train(
  STEAM.obj,
  model,
  n.tree,
  kernel,
  cv.folds,
  cv.repeats,
  trainval.ratio,
  train.folder.name,
  allowParallel,
  maxnweights = 5000
)
}
\arguments{
\item{STEAM.obj}{Object of class STEAM.Object}

\item{model}{A string specifying the model type. Options are \code{"rf"} (Random Forest), \code{"svm"} (Support Vector Machine), \code{"xgb"} (XGBoost), or \code{"multinom"} (Multinomial Logistic Regression).}

\item{n.tree}{Number of trees to use if \code{model = "rf"}.}

\item{kernel}{Kernel type for SVM. Options are \code{"radial"} or \code{"linear"}.}

\item{cv.folds}{Number of folds for cross-validation.}

\item{cv.repeats}{Number of repetitions for repeated cross-validation.}

\item{trainval.ratio}{Proportion of training data used for training in cross-validation.}

\item{train.folder.name}{Folder name where the trained model will be saved.}

\item{allowParallel}{Logical. If \code{TRUE}, allows parallel processing during model training.}

\item{maxnweights}{Maximum number of weights for multinomial logistic regression. Default is 5000.}
}
\description{
This function trains a classification model (e.g., Random Forest, SVM, XGBoost, or Multinomial Logistic Regression)
using the training data in a STEAM object. It performs repeated cross-validation and saves the trained model to a specified folder.
}
