% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nu.R
\name{nu}
\alias{nu}
\title{Helper Function for Ornstein Uhlenbeck Approximation}
\usage{
nu(x)
}
\arguments{
\item{x}{the numeric value you want to be evaluated.}
}
\value{
A single number.
}
\description{
Helper function for Ornstein Uhlenbeck analytic approximation to the family-wise error rate.
}
