% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_unc_nls.R
\name{run_unc_nls}
\alias{run_unc_nls}
\title{Run Unconstrained NLS Regression}
\usage{
run_unc_nls(lacorr, k1, k2, start.a = 0, start.b = 1, start.g = 10)
}
\arguments{
\item{lacorr}{local ancestry correlation; data frame with 3 columns: theta = recomb frac; corr = correlation; anc = ancestry components (e.g., 1_1, 1_2, 1_3, 2_2, 2_3, 3_3)}

\item{k1, k2}{ancestry component indices (number between 1 and K)}

\item{start.a}{starting value for intercept; default = 0}

\item{start.b}{starting value for slope; default = 1}

\item{start.g}{starting value for g; default = 10}
}
\value{
A single number indicating the estimated number of generations since admixture.
}
\description{
Estimate the number of generations since admixture (g) from the
observed local ancestry correlation curves using non-linear
least squares regression. More details in Grinde et al. (TBD).
}
\examples{
run_unc_nls(lacorr = example_corr, k1 = 1, k2 = 1)
run_unc_nls(lacorr = example_corr, k1 = 1, k2 = 2)

}
