% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSTEAM.R
\name{RunSTEAM}
\alias{RunSTEAM}
\title{Training and Evaluating Clustering performance with RunSTEAM()}
\usage{
RunSTEAM(
  matrix,
  coordinates,
  labels,
  train_ratio,
  n_size,
  seed,
  n_tree = 100,
  model = "rf",
  kernel = "radial"
)
}
\arguments{
\item{matrix}{gene expression matrix}

\item{coordinates}{spatial/image coordinates}

\item{labels}{cell annotations}

\item{train_ratio}{Fraction to set for training (if 0.8 is given, 80\% of data will be split for training. Default is 0.8)}

\item{n_size}{Neighbourhood Size for averaging}

\item{seed}{seed value.Default is 123}

\item{n_tree}{Number of trees}

\item{model}{rf or svm}

\item{kernel}{Default is Radial for svm}
}
\value{
Training model
}
\description{
Training and Evaluating Clustering performance with RunSTEAM()
}
