% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_correction.R
\name{safeRequirePackage}
\alias{safeRequirePackage}
\title{Safe Require Package}
\usage{
safeRequirePackage(package_name, install_message = NULL)
}
\arguments{
\item{package_name}{Name of the package}

\item{install_message}{Custom installation message}
}
\value{
TRUE if package is available
}
\description{
Safely requires a package with informative error message
}
