% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simstat_once.R
\name{simstat_once}
\alias{simstat_once}
\title{Simulate Admixture Mapping Test Statistics}
\usage{
simstat_once(m, K, as, bs, L)
}
\arguments{
\item{m}{the number of markers.}

\item{K}{the number of ancestral populations}

\item{as}{vector of scalars (length m) based on generations since admixture and distance between markers.}

\item{bs}{vector of scalars (length m) based on generations since admixture and distance between markers.}

\item{L}{matrix (K x K-1) that is square root of correlation matrix at single locus}
}
\value{
A single nummber indicating the largest simulated test statistic.
}
\description{
Simulate admixture mapping test statistics (one time)
from their joint asymptotic distribution under the null
(see Grinde et al., 2018).
}
\examples{
simstat_once(m = 2, K = 2, as = example_ab$a, bs = example_ab$b, L = example_L)

}
