% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STEAM.R
\docType{package}
\name{STEAM}
\alias{STEAM}
\alias{STEAM-package}
\title{STEAM: Significance Threshold Estimation for Admixture Mapping}
\description{
This package provides two functions for estimating genome-wide significance thresholds for
admixture mapping studies using two approaches developed by Grinde, Brown, Reiner,
Thornton, and Browning: get_thresh_analytic or get_thresh_statsim. We also provide a
function (get_g) which estimates the number of generations since admixture, which is
a required input parameter for both significance threshold estimation functions.
}
\section{STEAM Functions}{

The function get_thresh_analytic() estimates significance thresholds
using an analytic approximation to the family-wise error rate.

The function get_thresh_statsim() estimates significance thresholds by
simulating test statistics from their asymptotic distribution.

The function get_g() estimates the number of generations since admixture from
the observed pattern of local ancestry correlation.
}

