% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_correction.R
\name{validateFoldParameter}
\alias{validateFoldParameter}
\title{Validate Fold Parameter}
\usage{
validateFoldParameter(STEAM.obj, fold)
}
\arguments{
\item{STEAM.obj}{STEAM object containing fold data}

\item{fold}{Fold number to validate}
}
\value{
TRUE if valid, stops execution with error message if invalid
}
\description{
Validates fold parameter against available folds in STEAM object
}
