% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.split.R
\name{data.split}
\alias{data.split}
\title{Split STEAM object into training and testing sets}
\usage{
data.split(STEAM.obj, train.ratio)
}
\arguments{
\item{STEAM.obj}{A STEAM object with expression, labels, and spatial slots}

\item{train.ratio}{Numeric (0–1), proportion of data used for training}
}
\value{
Modified STEAM object with train/test subslots
}
\description{
Split STEAM object into training and testing sets
}
\keyword{internal}
