% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_expression.R
\name{feature_expression}
\alias{feature_expression}
\title{Feature Expression}
\usage{
feature_expression(steam_obj, feature_name, title = "Expression Across Layers")
}
\arguments{
\item{steam_obj}{STEAM Obj}

\item{feature_name}{Feature Name}

\item{title}{title}
}
\description{
Feature Expression
}
