% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_importance.R
\name{feature_importance}
\alias{feature_importance}
\title{Feature Importance}
\usage{
feature_importance(STEAM.obj, top_n = 10, title = "Top Features by Importance")
}
\arguments{
\item{STEAM.obj}{STEAM Object}

\item{top_n}{Top n}

\item{title}{plot title}
}
\description{
Feature Importance
}
