% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STEAM_Obj.R
\name{LoadSTEAM}
\alias{LoadSTEAM}
\title{STEAM object - Loading data for STEAM}
\usage{
LoadSTEAM(
  count_exp = NULL,
  spatial = NULL,
  labels = NULL,
  pca = NULL,
  umap = NULL,
  clusters = NULL,
  Seurat.obj = NULL,
  label.column = NULL,
  assay = "SCT"
)
}
\arguments{
\item{count_exp}{Count matrix/ Gene expression data}

\item{spatial}{Spatial Coordinates}

\item{labels}{Cluster info/annotations}

\item{pca}{PC embeddings (if using Seurat.obj)}

\item{umap}{UMAP embeddings (if using Seurat.obj)}

\item{clusters}{cluster labels (if using Seurat.obj)}

\item{Seurat.obj}{SEURAT Obj}

\item{label.column}{cluster labels}

\item{assay}{assay name (if using Seurat.obj)}
}
\value{
Object of class 'STEAM.Object'
}
\description{
STEAM object - Loading data for STEAM
}
\examples{
\dontrun{
  data("DLPFC")
  obj <- LoadSTEAM(count_exp = DLPFC$count_exp,
                   spatial = DLPFC$Spatial,
                   labels  = DLPFC$labels)
}
}
