% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.predict.R
\name{model.predict}
\alias{model.predict}
\title{Predict Labels and Evaluate Model Performance}
\usage{
model.predict(STEAM.obj)
}
\arguments{
\item{Object}{of class STEAM.Object}
}
\description{
This function takes a STEAM object, predicts labels for the test dataset using the trained model, and calculates performance metrics, including accuracy, precision, recall, F1-score, mean balanced accuracy, Adjusted Rand Index (ARI), and Kappa.
}
