% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood.avg.R
\name{neighborhood.avg}
\alias{neighborhood.avg}
\title{Perform Neighborhood Averaging}
\usage{
neighborhood.avg(STEAM.obj, n.size, is_train)
}
\arguments{
\item{STEAM.obj}{}

\item{n.size}{An integer specifying the neighborhood size (distance) for averaging.}

\item{is_train}{A logical value indicating whether to apply neighborhood averaging on the training dataset (\code{TRUE}) or the test dataset (\code{FALSE}).}
}
\description{
This function performs neighborhood averaging on the expression data in a STEAM object.
For each cell, it computes the average expression of neighboring cells within a specified spatial distance (\code{n.size}).
}
