% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_misclassified_cells.R
\name{plot_misclassified_cells}
\alias{plot_misclassified_cells}
\title{Plotting Misclassified labels}
\usage{
plot_misclassified_cells(steam_obj, label_colors = NULL)
}
\arguments{
\item{steam_obj}{STEAM Object}

\item{label_colors}{list of colors for plotting labels}
}
\description{
Plotting Misclassified labels
}
