% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_misclassified_cells.R
\name{plot_misclassified_cells}
\alias{plot_misclassified_cells}
\title{Plotting Misclassified labels}
\usage{
plot_misclassified_cells(steam_obj, coordinates = NULL)
}
\arguments{
\item{steam_obj}{STEAM Object}

\item{coordinates}{coordinates}
}
\description{
Plotting Misclassified labels
}
