% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSTEAM.R
\name{RunSTEAM}
\alias{RunSTEAM}
\title{Training and Evaluating Clustering performance with RunSTEAM()}
\usage{
RunSTEAM(
  STEAM.obj,
  train.ratio = 0.8,
  n.size = 5,
  seed = 123,
  cv.folds = 10,
  cv.repeats = 3,
  trainval.ratio = 0.8,
  model = "rf",
  n.tree = 500,
  kernel = "linear",
  train.folder.name = "train.out",
  allowParallel = FALSE
)
}
\arguments{
\item{STEAM.obj}{Object of class STEAM.Object from LoadSTEAM()}

\item{train.ratio}{training set ratio. Default is 0.8 (Splits 80\% data for training and 20\% for testing)}

\item{n.size}{Neighborhood size for neighborhood averaging}

\item{seed}{Seed Value. Default is 123}

\item{cv.folds}{#folds for Cross-validation}

\item{cv.repeats}{#repeats for repeated cross validation}

\item{trainval.ratio}{training set ratio for cross-validation. Default is 0.8 (Splits 80\% data for training and 20\% for validation)}

\item{model}{Currently available methods - rf (for random forest) , svm (for Support Vector Machine), xgb (for Extreme Gradient Boosting), multinom (for Penalized Multinomial Regression)}

\item{n.tree}{#trees (argument valid only for Random Forest). Deafult is 500}

\item{kernel}{Default is Linear (argument valid only for svm)}

\item{train.folder.name}{Training Output folder name}

\item{allowParallel}{Boolean to perform parallel processing, if resources are available. Default is FALSE}
}
\value{
S4 Object of class 'STEAM.Object' with all outputs
}
\description{
Training and Evaluating Clustering performance with RunSTEAM()
}
